[hipblaslt][CMS] Refactor validator to handle NGL and NLL  (#3279)
## Motivation

Previous implementation correctly handled the main loop of the body,
where the previous, current, and subsequent iterations were identical.
But did not check the no global load loop (NGL) and the no local load
loop (NLL).

In the NGL loop there are no GRs issued.**Hazards:**
1. Wrong value for nglshift (too small).

In the NLL loop there are no LR1s issued: **Hazard:**
1. Wrong value for nllshift (too small).
2. SWaits which guard LR0s but whose `dscnt` value accounts for LR1s.

## Technical Details

Moved from sliding window approach to instantiating multiple loops with
independent objects. Trading off performance of the sliding window
approach for substantially increased readability in handling the NGL and
NLL. Will come back to it in the future if there is a readable way to
implement efficiently.

## Test Plan

1. Existing tests
3. Add new tests

## Test Result

1. Caught a bug in 192x320x64 NT: #3287
2. New tests catch same issue and explain corner case

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>