[hipblaslt] Add option in CMS to set lgkmcnt to zero for NLL (#3311)
## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->
Add option in CMS schedule to set all lgkmcnt to zero for NLL.

Setting the `nglnllZeroDscnt` to True will enable the option, default
value is False.

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->

This is a quick workaround for some CMS schedules which have lgkmcnt
which depends on local reads for the next iteration. In the current
implementation, in those cases, the lgkmcnt for NLL may not be correct.

A proper fix which modifies a subset of the lgkmcnts will be added later
- some refactoring work required.

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->
Local testing pass, inspected generated asm to confirm macro supports
zeroing out lgkmcnts in NLL

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.