[MIOpen] MI355 3D Tunanet (solver selection heuristics) (#3154)
Note: this PR builds on #2868 and already incorporates the improved
iGEMM kernel suggested there. Otherwise the comparison would not be
fair, and we would not be testing the solver selection properly

## Motivation

This PR introduces Solver selection Heuristics (Tunanet) for 3D
convolutions on MI355. The architecture follows the Tunanet model used
for the 2D conv heuristics already in place.

As in the 2D conv case, this heuristics model takes preference over the
existing hardcoded WTI heuristics.

## Technical Details

* (Note files changed presently also list the changes made in #2868 as
that has not been merged yet)
* new model and metadata files for a 3D tunanet model for gfx950
* `Gfx942Model_3D` has been refactored into `TunaNet3DModel` where the
architecture/device name is now an argument (it will internally add the
"_3d" suffix to correctly find the correct paths etc).
* FYI: `TunaNet3DModel` is based on the Tunanet model used for 2D
heuristics. It is called in `PredictSolver()` in case `problem.Is3d()`
is True.
* I have also slightly refactored the flow internally so that the 3D
tunanet flow more closely follows (and reuses part of) the flow for 2D
tunanet.

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan
The tests follow the format introduced in #2868
I will run a selection of MIOpenDriver commands that have not been used
for training the model itself, i.e. a proper test set of 1000
commands/points (see attached [.txt
file](https://github.com/user-attachments/files/23771279/test_fdb_keys_random1000.drvcmds.txt)).
The performance will be compared to a build of the branch of PR #2868,
since this branchs should only differ from the develop branch in the
sense that it incorporates the improved heuristics.

The tests now differ, since we want MIOpen to also be responsible for
selecting the relevant solver (rather than defaulting to the iGEMM CK
solver as was the case in #2868)

### Env vars:
Env vars were set up such that the experimental branch was forced to
rely on the heuristics that we are testing. The Baseline branch was
either forced to use whatever fast mode was in place (i.e. for
comparison with currently implemented WTI for solver selection), or full
tuning (for the comparison to the theoretical maximum performance).

#### experimental + baseline heuristics:
```
  MIOPEN_FIND_ENFORCE: "1" # NONE/(1): No change in the default behavior.
  MIOPEN_FIND_MODE: "2" # (fast find): Checks FindDb for an entry. If there’s a FindDb hit, it uses that entry. 
    #If there’s a miss, it uses the immediate mode fallback (= heuristics we are testing).
  MIOPEN_USER_DB_PATH: A directory unique for all benchmarks, and cleared before running the shapes
```
#### baseline full tuning:
Note that in order to save quite some time running the benchmark, we
disabled Naive solvers.
```
    MIOPEN_DEBUG_CONV_GEMM: "1"
    MIOPEN_DEBUG_CONV_DIRECT: "1"
    MIOPEN_DEBUG_CONV_WINOGRAD: "1"
    MIOPEN_DEBUG_CONV_IMPLICIT_GEMM: "1"
    MIOPEN_DEBUG_CONV_FFT: "1"
    MIOPEN_DEBUG_CONV_DIRECT: "0" # still skip the naive solvers
    MIOPEN_FIND_ENFORCE: "3"
    MIOPEN_FIND_MODE: "1"
    MIOPEN_SYSTEM_DB_PATH: "/tmp/miopen_perf_test"
```
## Test Result

### Compared to current heuristics
<img width="1000" height="500" alt="image"
src="https://github.com/user-attachments/assets/92b8a463-f542-44a9-bf3d-ff92a1c8f304"
/>


#### summary
on average (wrt current WTI heuristics):

* BWD: ~30% faster
* FWD: ~820% faster
* WRW: ~740% faster


### Compared to full tuning
<img width="1000" height="500" alt="image"
src="https://github.com/user-attachments/assets/9c3887ca-bf9c-432e-a803-7c51cd7118a0"
/>


#### summary
Note that when the iGEMM solver is most optimal, these numbers are
compounded by the perf metrics from #2868, so I also mention those
numbers here
on average (wrt full tuning):

* BWD: 90% of optimal performance (94% for iGEMM heuristics only)
* FWD: 89% of optimal performance (93% for iGEMM heuristics only)
* WRW: 87% of optimal performance (89% for iGEMM heuristics only)


## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.