[hipblaslt] [CMS] Refactor CMS schedule registering to use decorator (#3325)
## Motivation

This allows each CMS kernel to be added independently of the others and
removes the large if-else chain. This will let us cleanly refactor later
to avoid any future merge conflicts with this file when adding new
schedules.

## Technical Details

1. Add global list to programmatically append schedules to.
2. Wrap each implementation in a decorator which adds it to the list and
performs the checks that used to performed by the chained if-else.
3. Rewrite `has_custom_schedule` to iterate through list and find
matching schedule (if any).

## Test Plan

Run existing tests.

## Test Result

Passing

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>