[hipblaslt [CMS] Move setting of LocalRead.needed_by out of initializer (#3382)
## Motivation

Move the setting of LocalRead.needed_by from initializer into separate
function to as setup for allowing validation of LocalReads to be done
separately from validation of GlobalReads.

Depends on #3361.

## Technical Details

Move to helper function that is called from validation passes that
require it, same as `apply_barriers()`.

## Test Plan

Run existing tests.

## Test Result

Passes.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>