[hipblaslt] [CMS] Common test class for validator testing (#3362)
## Motivation

Refactor all of the validator tests to use a common test class.

Need it as a future work when changing the interface for the validators
to take in a timeline object rather than each creating one internally.

No functional changes to the validator passes.

Depends on #3361.

## Technical Details

1. Create common class to have central location for interfacing with a
particular validator pass.
2. Split several tests that included multiple tests per method into 1
test per method.
3. Rewrite how schedule is defined for certain tests to use a unified
approach for all tests.

## Test Plan

Run existing tests.

## Test Result

Passing.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>