[hipDNN] [ALMIOPEN-490] Untangling test_sdk and plugin_sdk (#3458)
## Motivation
The goal of this PR is to untangle some dependencies between sdk,
test_sdk, and plugin_sdk code. The goal after this PR is to migrate the
contents of sdk/plugins (except for the flatbuffer wrappers) to a
separate plugin_sdk library, which has already been created and
integrated with TheRock.

## Technical Details
- Moved Mock*.hpp to test_sdk, updated namespaces and references
- Modified the flatbuffer wrappers to throw regular exceptions rather
than plugin exceptions
- Split out some flatbuffer graph generation code between test_sdk and
plugin_sdk versions

## Test Plan
- Run all unit tests
- Run all integration tests
- check target passes
- pre-commit passes

## Test Result
- [x] All integration, unit, and check targets passing

## Submission Checklist
- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.
- [ ] Notify the IREE folks when this lands