[MIOpen] Bugfix Use uint64_t instead of long for offsets (#3381)
## Motivation

When porting OpenCL kernels the `ulong` datatype was often ported to
`unsigned long`. This is is problematic since the size of long on
windows is 32 bits. This MR replaces the uses of long with `uint64_t`.

refs: https://github.com/ROCm/rocm-libraries/pull/3364


## Test Plan

The tests should be extended in a follow up MR to cover these cases as
well.

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.